Refactor the dataset highlighting code to track state and improve performance
Tracking state will be required for future feature addition that will
need to interact with the highlighting code in a deterministic manner.

The cleanup required to track state allows for changes in the way DOM
elements are located which leads to fewer DOM searches.  This should
improve performance, possibily significantly.