Squash some tooltip bugs
Some recent changes require d.substr to become d.name.substr.

Additionally, get rid of the need to do a complicated DOM query to
remove the tooltip since it can be accessed directly.  This allows the
removal of the broken id name creation which could fail dependening on
the characters included in the dataset name.