Update jschart to be more dynamic when hiding and unhiding datasets
The charts will now update the domains depending on the visible
datasets and stacked charts will now be restacked when a dataset is
hidden or unhidden.

Also includes some bug fixes like properly hiding and unhiding single
point datasets when they are individually toggled.

The methodology which the stacked mean is calculated has been updated
to be more easily performed.

The code used to determine the current axis domains has been
consolidated into a single function for simplicity and consistency.