Update jschart to automatically sort the datasets in descending order based on their mean by default
This change ensure that the datasets that are most likely to be
relevant to the user show up early in the dataset list, which is
extremely valuable for very large dataset counts.

If this behavior is not desired, it can be overriden by adding
'sort_datasets: false' to the list of supplied options.  This will
fallback to the original behavior where datasets are presented in load
order based on their specification to create_graph.  This behavior is
ideal when chart consistency is desired (ie. dataset order, dataset
color, etc.)