Update jschart to improve maintainability and performance
Move a common block of code into a function to improve maintainability
and then optimize it.

This block of code updates the chart elements in reaction to dynamic
actions that the user takes so the work it must do scales with the
number of loaded datasets.  The first optimization is to avoid using
DOM searches to locate these elements.  Next, only operate on visible
datasets; updates to hidden datasets can be deferred until the
datasets are visible again on a subsequent call.

A third optimization is that when toggling a dataset's hidden status
the mouse_highlight_function call only needs to be made when the
toggle was initiated from a mouse event so allow the call to be
skipped when a threshold is being applied.