fs-version-002: New Versioning Mechanism with inotify
This PR will do the following changes:

1. pbench-move-results will create a new versioning
mechanism on the server side to deal with the
incompatibility: old clients will remain unchanged
and the data will be processed exactly as they are
processed today; new clients will use the new
mechanism.

2. pbench-server-prep-version-001 is a new script:
it will scour the reception directory for xxx.tar.xz.md5
files. For each file, it will check the md5 sum and, if
successful, it will copy the tarball and its md5 to the
archive directory and create the link in the watch state
directories. That will get the ball rolling on the rest
of the processing. Errors like missing tarball, missing
md5, duplicate names, are detected and handled
appropriately (mostly by quarantining the tarball).

3. This PR will change the archive directory structure
to make it deeper and not as wide, by using the MD5
string of the tarball to implement an (on-disk) trie
(like git, or gluster): e.g. the MD5 sum
7341243af6d9af1025cbec71023dbff7 will generate a pathname
of 73/41/7341243af6d9af1025cbec71023dbff7 to a directory
that will hold the unpacked "container" tarball.