Refactor tool-scripts to use one script
Addresses issue #1459 following "Proposal 1".

This is a fairly invasive and major change to the tool sub-system.  The
pbench-agent no longer tries to install **ANY** RPMs via yum.  The
interface to the `pbench-start/stop/postprocess-tools` has changed, and
changes to various interfaces and behaviors of individual tools have
also been made.

**Behavioral Changes**

The pbench-agent no longer tries to install *ANY* RPMs via `yum`.

This is significant in that all of the benchmark scripts now only
*check* that their required RPMs are installed, and all the tools
during registration only perform a check that their requirements are
satisfied.

Again, no RPMs are installed by the pbench-agent code.

All of the tool datalogs have been enhanced to check for their required
environment and issue error messages describing what is missing for them
to operate properly.

One major behavior change to the `systemtap` tool is worth noting.
The systemtap script argument is now the "compiled" systemtap
script. The user is now responsible for making sure the provided
script is available to the systemtap tool on all the hosts it is
registered.

**Interface Changes**

We are removing knowledge of how `bench-scripts` handle "iterations" and
"samples" so that the tools are only given a directory to work with,
`${tool_output_dir}`, and no longer use `${pbench_tmp}`.  All that they
need to do with files are managed in the `${tool_output_dir}`.

This then allowed us to remove the `--iteration` and `--group`
parameters from all the tools, and in turn, remove the `--iteration`
parameter from the `pbench-(start|stop|postprocess)-tools` commands.
This prevents knowledge of how benchmark scripts work from leaking into
the various tool implementations.

**Tool Script Changes**

We now have two scripts that implement the methods for running tools:

 * `base-tool`
    * a merger of all the previous tool script interfaces
    * with PR #1248, this will be replaced by Tool Meister code
    * all tools now provide a proper `${tool}-datalog` implementation,
      along with `${tool}-postprocess` and `${tool}-stop-postprocess`
      scripts where appropriate
    * the `haproxy-ocp`, `oc`, and `pprof` tools now have a Python 3
      based `${tool}-datalog`
 * `external-data-source`
    * dedicated to just handling for the no-op external data source

The `user-tool` interfaces just provide an arbitrary path to scripts
that should behave the same way as our existing `*-datalog`,
`*-stop-postprocess`, and `*-postprocess` scripts.  The `user-tool`
start script is passed the `${tool_output_dir}` and the interval
value.  The stop and post-process scripts are just passed the
`${tool_output_dir}`.

**Tool Post-Processing**

We have added the notion of a "stop-postprocess" script so that when
tools are stopped (i.e. when `pbench-stop-tools` is invoked), any
post-processing required to be performed on the host on which the
tool is running can be executed.  This allows us to run all the
normal post-procesing of data to generate charts and graphs during
the `pbench-postprocess-tools` phase only on the local controller, or
at a later time.

**Unit Tests**

We have added unit tests for all the tools in the `tool-scripts`
directory, and in the `tool-scripts/datalog` directory.

**A Guide to Reviewing the Changes**

 * `agent/base`

   Changed to only check for RPMs, installation removed; we also now
   have a single `_PBENCH_UNIT_TESTS` environment variable in use

 * `agent/util-scripts`

   Changes to remove the use of `--iteration` when invoke tools;
   accommodate new list of available tools

 * `agent/bench-scripts`

   Changes to remove the use of `--iteration` when invoking tools

 * `agent/tool-scripts`

   Except for the "external data source", all 37 tool files are now
   symlinks to `base-tool`, with it containing all the tool iterface
   code and how the `${tool}-datalog`, `${tool}-stop-postprocess`,
   `${tool}-postprocess` are invoked

 * `agent/tool-scripts/datalog`

   All tools now have a `${tool}-datalog`; the `pidstat-convert` script
   now lives in this directory, moved from `../postprocess`; we share
   a common `File-Capture-datalog` for `proc-interrupts`,
   `proc-sched_debug`, and `proc-vmstat`

 * `agent/tool-scripts/postprocess`

   No-op post-processing scripts have been removed (`dm-cache`,
   `systemtap`, `tcpdump`, and `turbostat`); we have added
   `${tool}-stop-postprocess` scripts where necessary (`blktrace`,
   `cpuacct`, `kvmtrace`, `mpstat`, and `perf`); a number of the
   post-processing scripts have been fixed to accommodate the changes to
   their `-datalog` counter parts.  The `pcp-postprocess` script has
   been changed to use `pcp2csv`.

Lastly, we now use `systemd-detect-virt` where applicable.