Add a pbench namespace
Add a pbench namespace for the python bits for pbench. Using
one namespace allows us to share code between server and agent
and will cut down on the duplication that we do.

A side benefit of this is also allows pbench easier to build
and distribute via pip and RPM.

Most of the changes in these files are a straight move from
server/lib/pbench, lib/configtools, server/lib/s3backup since
we are consolidating everything in once place.

Unittests were updated so that we could still run
unittests when the code was moved around.

Signed-off-by: Charles Short <chucks@redhat.com>