Fix slow unit tests
Apparently, when we use `files` in `setup.cfg`, that causes a separate
`bin/pbench-config` module (auto-generated) to be created, which loads
the file `bin/pbench-config` in the tree and invokes `compile` on it
each time.

This change avoids the compiled version so that all `pbench-config`
commands run normally (still slow, but much faster than being compiled
each time).