Add context decorator
Whenever a Click command is executed, a Context object is created which
holds state for this particular invocation. It remembers parsed parameters,
what command created it, which resources need to be cleaned up at the end
of the function, and so forth. It can also optionally hold an application-defined object.

Create an empty context object so we can maintain the state of the
agent.

Signed-off-by: Charles Short <chucks@redhat.com>
(cherry picked from commit 75ba130355c45f30faf6ce5ce65d8fdb178193db)